feat: export bundled types for TypeScript
