+++ test/test.index.ts
@@ -1,0 +1,26 @@
+// Copyright 2020 Google LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import {describe, it} from 'mocha';
+import {Auth, Common} from '../src';
+
+/* eslint-disable @typescript-eslint/no-unused-vars */
+
+describe(__filename, () => {
+  it('should export bundled package interfaces', () => {
+    let p: Auth.OAuth2Client;
+    let q: Auth.Credentials;
+    let t: Common.StreamMethodOptions;
+    let v: Common.SchemaResource;
+  });
+});
